Added variable to the StameMachinePlayer
Added an exported variable to the StateMachinePlayer so developers can specify the node passed to the state functions.  By default the node that is parent of the StateMachinePlayer is passed into the state functions.  This node will be passed instead if a node other then the parent node needs to be managed by your state code.